refactor: TestUser 분리
